storage: cleanup variables and reorganize
