+++ src/storage/mod.rs
@@ -1,12 +1,17 @@
 pub mod btree;
 pub mod constants;
+pub mod context;
 pub mod cursor;
 pub mod error;
 pub mod page;
 pub mod pager;
 
 pub use btree::Tree;
+pub use context::{AccessContext, AccessMode};
 pub use cursor::Cursor;
 pub use error::{Result, StorageError};
-pub use page::{MetaPage, Page, PageFlags, PageView, PageViewMut, TablePage};
-pub use pager::{AccessContext, FlushGuard, Pager};
+pub use page::{
+    MetaPage, Mutation, MutationScope, Page, PageFlags, PageView, PageViewMut,
+    TablePage,
+};
+pub use pager::{FlushGuard, Pager};
